+++ test/src/terminal_test.dart
@@ -2884,8 +2884,8 @@ void main() {
       '\x1b[?6c',
       '\x1b[?6c',
       '\x1b[?6c',
-      '\x1b[>0;50000;1c',
-      '\x1b[>0;50000;1c',
+      '\x1b[>0;50300;1c',
+      '\x1b[>0;50300;1c',
     ]);
   });
 
@@ -2946,7 +2946,7 @@ void main() {
     terminal.write('\x1b[>0q');
 
     expect(output, [
-      '\x1bP>|xterm2 5.0.0\x1b\\',
+      '\x1bP>|xterm2 5.3.0\x1b\\',
       '\x1bP>|lumide-term 1.0\x1b\\',
     ]);
   });